rename to more accurate -> storage
the storage can be basically split into InMemoryPool and PersistentPool
